Implemented a findBestPurchase function in Economy and added depth limits to GraphMap BFS search

+++ SFMLSpaceGame/ItemPrice.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef unsigned int Price;
+typedef int64_t Price;
